Make chat.java thread safe. 1st try
Still not sure if the reading of values from the map must be included
in a mutual exclusion zone. Some users could be removed while a
message is trying to be sent to them. However, that could make longer
any handling with users if the call to send the message is blocking.