handle a sendmessage in a thread per user
A completionSErvice is used with and executor of one
single thread to manage the request send for a given
user. The completionService is created when the user
is added to a chat and shutdown when the user is
removed from the chat.

The test sometimes fail because the order is not the
correct one.
Still pending to change the other notifications, when
a user is added to the chat or a user leave a chat.