handle a notifications new an exist in a thread per user
A completionSErvice is used with and executor of one
single thread to manage the request send for a given
user. The completionService is created when the user
is added to a chat and shutdown when the user is
removed from the chat.

The test sometimes fail because the order is not the
correct one.
The second test fails as well. check it.