Change way of check the user has sent the msg
Previouly an array of string was used. The name of the user was
set in a position. But because the name cannot been in order, it is
enough to know if the message was sent, so a boolean does the work.
Another option is to use a set, an compare a template set with the
returned set.